Fix: mismatch between protein and DNA ref coordinate for mouse
Drop any ref DNA sequence not matching with protein ones and replace them using backtranslation

This fix avoid to use reference coordinates translation in any case there is a mismatch between the reference genome used to retrieve the coordinates (Uniprot Coordinates API) and the one used to retrieve the DNA sequences (BGreference and BGdata).

Currently, we hypothesise there is such a mismatch in the case of "Mus musculus" with Uniprot using mm39 and BGreference having only mm10 available.

We should update BGreference, use mm39 and test if the mismatch persist.

This is a temporary fix that it is required in any case the coordinates obtained from Uniprot Coordinates API are not matching with the reference genome used to retrieve the sequences.